refactor to use command extension
